[eas-cli] rename branch:publish to update:publish
